# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "",
      "description": ""
    }
  ],
  "schemes": [],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/users/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Adding new user.",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "doruk"
                },
                "password": {
                  "type": "string",
                  "example": "dodo1234"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Logging in.",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "doruk"
                },
                "password": {
                  "type": "string",
                  "example": "dodo1234"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "New password",
            "schema": {
              "type": "object",
              "properties": {
                "newPassword": {
                  "type": "string",
                  "example": "doruk"
                }
              },
              "required": [
                "newPassword"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/whoami": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/addCredit": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Adding credit",
            "schema": {
              "type": "object",
              "properties": {
                "credit": {
                  "type": "number",
                  "example": 50
                }
              },
              "required": [
                "credit"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/auctions/": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Adding new auction.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "doruk"
                },
                "startingPrice": {
                  "type": "number",
                  "example": 50
                },
                "closingPrice": {
                  "type": "number",
                  "example": 60
                }
              },
              "required": [
                "name",
                "startingPrice",
                "closingPrice"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Auctions"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/auctions/{auctionId}": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Auctions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Auctions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "parameter_name",
            "in": "body",
            "description": "Put auction",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "dorukChanged"
                },
                "startingPrice": {
                  "type": "number",
                  "example": 55
                },
                "closingPrice": {
                  "type": "number",
                  "example": 65
                }
              },
              "required": [
                "name",
                "startingPrice",
                "closingPrice"
              ]
            }
          }
        ],
        "responses": {}
      }
    },
    "/offers/{auctionId}": {
      "post": {
        "tags": [
          "Offers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "parameter_name",
            "in": "body",
            "description": "Post offer",
            "schema": {
              "type": "object",
              "properties": {
                "price": {
                  "type": "number",
                  "example": 70
                }
              },
              "required": [
                "price"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/offers/{offerId}": {
      "put": {
        "tags": [
          "Offers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "parameter_name",
            "in": "body",
            "description": "Put offer",
            "schema": {
              "type": "object",
              "properties": {
                "price": {
                  "type": "number",
                  "example": 71
                }
              },
              "required": [
                "price"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Offers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/offers/": {
      "get": {
        "tags": [
          "Offers"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/notifications/": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Post notification",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "temp notification."
                }
              },
              "required": [
                "message"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Notifications"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/notifications/{notificationId}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "description": "",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Notifications"
        ],
        "description": "",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Post notification",
            "schema": {
              "type": "object",
              "properties": {
                "isRead": {
                  "type": "boolean",
                  "example": true
                }
              },
              "required": [
                "isRead"
              ]
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "definitions": {}
}